+++ tests/testthat/test-5mixed_kernel.R
@@ -10,6 +10,20 @@ test_that("Mixed kernels", {
 
 })
 
+test_that("Mixed kernels multiple Hurst", {
+
+  mod1 <- kernL(stack.loss ~ ., stackloss,
+                model = list(kernel = c("FBM,0.1", "Canonical", "FBM,0.1")))
+  mod2 <- kernL(stack.loss ~ ., stackloss,
+                model = list(kernel = c("FBM", "Canonical", "FBM"), Hurst = 0.1))
+  expect_equivalent(mod1$Hl, mod2$Hl)
+  expect_warning(
+    kernL(stack.loss ~ ., stackloss,
+          model = list(kernel = c("FBM,0.9", "Canonical", "FBM,0.9"), Hurst = 0.1))
+  )
+
+})
+
 test_that("Warn when using one.lam = TRUE", {
 
   expect_warning(